# exec: consolidate all Append{...} methods into one