--- a/pkg/sql/exec/hashjoiner.go
+++ b/pkg/sql/exec/hashjoiner.go
@@ -467,14 +467,16 @@ func makeHashTable(
 // output columns.
 func (ht *hashTable) loadBatch(batch coldata.Batch) {
 	batchSize := batch.Length()
-	sel := batch.Selection()
-
 	for i, colIdx := range ht.valCols {
-		if sel != nil {
-			ht.vals[i].AppendWithSel(batch.ColVec(int(colIdx)), sel, batchSize, ht.valTypes[i], ht.size)
-		} else {
-			ht.vals[i].Append(batch.ColVec(int(colIdx)), ht.valTypes[i], ht.size, batchSize)
-		}
+		ht.vals[i].Append(
+			coldata.AppendArgs{
+				ColType:   ht.valTypes[i],
+				Src:       batch.ColVec(int(colIdx)),
+				Sel:       batch.Selection(),
+				DestIdx:   ht.size,
+				SrcEndIdx: batchSize,
+			},
+		)
 	}
 
 	ht.size += uint64(batchSize)