--- a/pkg/sql/exec/sorttopk.go
+++ b/pkg/sql/exec/sorttopk.go
@@ -162,9 +162,7 @@ func (t *topKSorter) spool(ctx context.Context) {
 					// TODO(solon): Make this copy more efficient, perhaps by adding a
 					// copy method to the vecComparator interface. This would avoid
 					// needing to switch on the column type every time.
-					t.topK.ColVec(j).AppendSlice(
-						inputBatch.ColVec(j), t.inputTypes[j], uint64(maxIdx), uint16(idx), uint16(idx)+1)
-
+					t.topK.ColVec(j).CopyAt(inputBatch.ColVec(j), uint64(maxIdx), uint64(idx), uint64(idx+1), t.inputTypes[j])
 				}
 				heap.Fix(t, 0)
 			}