This makes the code much easier to work with.

```
name                    old time/op    new time/op    delta
Append/SimpleAppend-8      199ns ± 4%     190ns ± 1%   -4.51%  (p=0.000 n=9+9)
Append/AppendWithSel-8    2.45µs ± 1%    2.13µs ± 2%  -13.04%  (p=0.000 n=9+9)

name                    old speed      new speed      delta
Append/SimpleAppend-8   41.0GB/s ± 4%  42.9GB/s ± 2%   +4.64%  (p=0.000 n=9+9)
Append/AppendWithSel-8  3.34GB/s ± 1%  3.84GB/s ± 2%  +14.99%  (p=0.000 n=9+9)
```

I'm not sure why this is faster in the SimpleAppend case (it's possible
it's just noise given the variance), but the AppendWithSel case benefits
from less index arithmetic and bounds checks.

Release note: None