@asubiotto 2019-06-24
Removed the behavior that `DestIdx` defaults to the end of the destination slice given that some operators explicitly pass in `0` 🤷‍♂ . I think it's probably safest and least surprising to keep it this way. Maybe we want to do the same with `SrcEndIdx`? It's explicitly specified in every callsite anyway.