@solongordon 2019-06-24
I wondered about that. Yeah, I would lean towards making `SrcEndIdx` required as well for least surprise.